Reuse some of the tests written for BulkUpdate
